refactor: side menu button
