+++ app/(media)/_components/AppBar/MediaListButton.tsx
@@ -1,0 +1,22 @@
+import { IconButton, SxProps, Theme } from "@mui/material";
+import React from "react";
+import { GridViewRounded as GridViewRoundedIcon } from "@mui/icons-material";
+import Link from "next/link";
+
+const buttonSx: SxProps<Theme> = {
+  backgroundColor: (theme) => theme.palette.grey["800"],
+};
+
+export default function MediaListButton() {
+  return (
+    <IconButton
+      component={Link}
+      href="/"
+      color="inherit"
+      edge="start"
+      sx={buttonSx}
+    >
+      <GridViewRoundedIcon />
+    </IconButton>
+  );
+}
